+++ src/indexer/romanz_electrsd/versions.rs
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+/// Human-readable name of [`RomanzElectrsD`](crate::indexer::romanz_electrsd::RomanzElectrsD).
+#[allow(unused)]
+pub(crate) const ROMANZ_ELECTRS_NAME: &str = "RomanzElectrsD";
+
+/// Binary name of [`RomanzElectrsD`](crate::indexer::romanz_electrsd::RomanzElectrsD).
+#[allow(unused)]
+pub(crate) const ROMANZ_ELECTRS_BIN_NAME: &str = "romanz_electrs_v_0_12_0";
+
+/// `romanz/electrs` v0.12.0.
+#[allow(unused)]
+pub(crate) const ROMANZ_ELECTRS_VERSION: &str = "0.12.0";