+++ src/lib.rs
@@ -15,18 +15,18 @@
 //! | Node    | [`Floresta`]            | `v0.9.1`  | `florestad`           |                        |
 //! | Node    | [`utreexod`]            | `v0.6.0`  | `utreexod`            |                        |
 //! |         |                         |           |                       |                        |
-//! | Indexer | [`romanz/electrs`]      | `v0.12.0` | `electrs`             |                        |
-//! | Indexer | [`mempool/electrs`]     | `v3.3.0`  | `mempool_electrs`     | Unsupported on Windows |
 //! | Indexer | [`Blockstream/electrs`] | `4b1a018` | `blockstream_electrs` | Unsupported on Windows |
 //! | Indexer | [`ElectrumX`]           | `v1.20.0` | `electrumx`           | Needs Python 3.10      |
+//! | Indexer | [`mempool/electrs`]     | `v3.3.0`  | `mempool_electrs`     | Unsupported on Windows |
+//! | Indexer | [`romanz/electrs`]      | `v0.12.0` | `romanz_electrs`      |                        |
 //!
 //! [`Bitcoin Core`]: <https://github.com/bitcoin/bitcoin>
 //! [`btcd`]: <https://github.com/btcsuite/btcd>
 //! [`Floresta`]: <https://github.com/getfloresta/Floresta>
 //! [`utreexod`]: <https://github.com/utreexo/utreexod>
-//! [`romanz/electrs`]: <https://github.com/romanz/electrs>
-//! [`mempool/electrs`]: <https://github.com/mempool/electrs>
 //! [`Blockstream/electrs`]: <https://github.com/Blockstream/electrs>
+//! [`mempool/electrs`]: <https://github.com/mempool/electrs>
+//! [`romanz/electrs`]: <https://github.com/romanz/electrs>
 //! [`ElectrumX`]: <https://github.com/spesmilo/electrumx>
 //! [`electrumx`]: <https://github.com/spesmilo/electrumx>
 //! [`Indexer`]: crate::indexer::Indexer