+++ src/node/mod.rs
@@ -61,10 +61,10 @@ pub(crate) const MIN_PRUNE_TARGET_MIB: u64 = 550;
     feature = "bitcoind",
     feature = "btcd",
     feature = "utreexod",
-    feature = "electrs",
+    feature = "blockstream_electrs",
     feature = "electrumx",
     feature = "mempool_electrs",
-    feature = "blockstream_electrs"
+    feature = "romanz_electrs"
 ))]
 pub(crate) const RPC_COOKIE_FILE_NAME: &str = ".cookie";